Feedback from the reviewer =================
What did the author do particularly well?
----------------------------------
a neat little story form an author who apparently does not have command English as the first language. the spellings are good, and so is much of the grammar and idiom.

What could the author improve on?
------------------------------
the author is in trouble with syntax, and use of prepositions. some of the 'action scenes' - the ones where the magic happens are described too suddenly and abruptly. a little more description would improve the narration. i see the author rewriting the story for another round of reviewing.